cloudNOU/actuate/requirements.txt
paho-mqtt
kafka-python
influxdb-client